Refactor. Remove Update for event selector
