+++ app/views/layouts/admin.html.erb
@@ -4,6 +4,7 @@
   <%= stylesheet_link_tag 'layouts/admin', media: 'all', 'data-turbolinks-track': 'reload' %>
   <%= javascript_pack_tag 'server_rendering' %>
   <%= javascript_pack_tag 'take_attendance_application' %>
+  <%= javascript_pack_tag 'attendance_event_submission_answer' %>
 <% end %>
 
 <% content_for :content do %>
